Fixes y fluent validation acabada 100%
