+++ RealEstateApp.Core.Application/Features/SalesTypes/Queries/List/ListSalesTypesQuery.cs
